poros_backend/app/requirements.txt
## Poros Protocol Backend Dependencies

# Core framework
fastapi>=0.100.0
uvicorn[standard]>=0.23.0

# A2A Protocol SDK (Google's official SDK)
a2a-sdk[all]>=0.3.0

# Database
sqlmodel>=0.0.18
sqlalchemy>=2.0.0

# HTTP client for calling agents
httpx>=0.24.0

# Auth
python-jose[cryptography]>=3.3.0
passlib>=1.7.4
bcrypt==4.0.1
python-multipart>=0.0.6

# Environment
pydantic-settings>=2.0.0
python-dotenv>=1.0.0

# Utilities
python-dateutil>=2.8.2
